backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.5.0
asgiref==3.8.1
attrs==24.2.0
boto3==1.35.22
botocore==1.35.22
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cohere==5.9.4
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fastapi==0.115.6
fastavro==1.9.7
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.0
idna==3.7
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.12
langchain-community==0.3.12
langchain-core==0.3.25
langchain-openai==0.2.12
langchain-text-splitters==0.3.3
langserve==0.3.0
langsmith==0.2.3
Markdown==3.6
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.57.4
orjson==3.10.12
packaging==24.1
parameterized==0.9.0
propcache==0.2.1
psycopg2==2.9.9
pydantic==2.9.2
pydantic-settings==2.7.0
pydantic_core==2.23.4
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.1
sse-starlette==1.8.2
starlette==0.41.3
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.20.0
tqdm==4.66.5
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.33.0
yarl==1.18.3
